Update backend_wx.py to fix missing coordinates in figure toolbar
The static text control used to display coordinates is currently not given a size and is allocated  minimal width on toolbar. Realize() hence coordinates are not visible.

To fix add an initial size argument (here I've used 120 which should be sufficient for most font sizes).

An alternative fix is to call realise() when updating the message in set_message() however this resizes the entire toolbar and causes flickering on my system (win 11) at least.